requirements.txt
azure.functions
pyodbc==5.1.0